requirements.txt
slack-bolt>=1.18.0
apscheduler>=3.10.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
